module_14_3

+++ module_14/module_14_2.py
@@ -4,8 +4,9 @@ import os
 import sqlite3
 from pathlib import Path
 
-os.remove("not_telegram.db")
-Path("not_telegram.db").touch()
+if os.path.isfile("not_telegram.db"):
+    os.remove("not_telegram.db")
+    Path("not_telegram.db").touch()
 
 connection = sqlite3.connect("not_telegram.db")
 cursor = connection.cursor()
@@ -31,7 +32,7 @@ users = cursor.fetchall()
 for u in users:
     id, name, email, age, balance = u
     if id % 2:
-        cursor.execute("UPDATE Users SET balance = ? WHERE id = ?",(balance + 500, id))
+        cursor.execute("UPDATE Users SET balance = ? WHERE id = ?",(500, id))
 connection.commit()
 # 3
 for id in range(1,10,3):
@@ -39,6 +40,13 @@ for id in range(1,10,3):
 
 cursor.execute("DELETE FROM Users WHERE id = ?",(6,))
 connection.commit()
+
+# 4
+print("#4")
+cursor.execute("SELECT * FROM Users WHERE age != 60")
+for row in cursor.fetchall():
+    id, username, email, age, balance = row
+    print(f"Имя: {username} | Почта: {email} | Возраст: {age} | Баланс: {balance}")
 
 cursor.execute("SELECT COUNT(*) FROM Users")
 total_users = cursor.fetchone()[0]
